requirements.txt
networkx==2.4
numpy==1.16.0
tqdm==4.46.1
pandas==1.0.5
pytorch==1.6.0
scipy==1.3.3
optuna==2.7.0
sklearn-pandas==1.7.0
matplotlib==3.1.3
